Move anchor server to pkg/anchor/server.go.
